Move tck files
