+++ packages/graphql/tests/tck/interfaces.test.ts
@@ -20,9 +20,9 @@
 import { Neo4jGraphQLAuthJWTPlugin } from "@neo4j/graphql-plugin-auth";
 import { gql } from "apollo-server";
 import type { DocumentNode } from "graphql";
-import { Neo4jGraphQL } from "../../../src";
-import { createJwtRequest } from "../../utils/create-jwt-request";
-import { formatCypher, translateQuery, formatParams } from "../utils/tck-test-utils";
+import { Neo4jGraphQL } from "../../src";
+import { createJwtRequest } from "../utils/create-jwt-request";
+import { formatCypher, translateQuery, formatParams } from "./utils/tck-test-utils";
 
 describe("Interfaces tests", () => {
     const secret = "the-secret";
